+++ crates/ruma-client-api/src/session/login.rs
@@ -417,8 +417,8 @@ pub mod v3 {
                 .unwrap(),
                 LoginInfo::Password(login)
             );
-            assert_matches!(login.identifier, Some(UserIdentifier::UserIdOrLocalpart(user)));
-            assert_eq!(user, "cheeky_monkey");
+            assert_matches!(login.identifier, Some(UserIdentifier::Matrix(id)));
+            assert_eq!(id.user, "cheeky_monkey");
             assert_eq!(login.password, "ilovebananas");
 
             assert_matches!(
@@ -443,7 +443,7 @@ pub mod v3 {
             use serde_json::Value as JsonValue;
 
             use super::{LoginInfo, Password, Request, Token};
-            use crate::uiaa::UserIdentifier;
+            use crate::uiaa::{EmailUserIdentifier, UserIdentifier};
 
             let supported = SupportedVersions {
                 versions: [MatrixVersion::V1_1].into(),
@@ -476,9 +476,9 @@ pub mod v3 {
             let req: http::Request<Vec<u8>> = Request {
                 #[allow(deprecated)]
                 login_info: LoginInfo::Password(Password {
-                    identifier: Some(UserIdentifier::Email {
-                        address: "hello@example.com".to_owned(),
-                    }),
+                    identifier: Some(UserIdentifier::Email(EmailUserIdentifier::new(
+                        "hello@example.com".to_owned(),
+                    ))),
                     password: "deadbeef".to_owned(),
                     user: None,
                     address: None,