+++ crates/ruma-client-api/src/uiaa/auth_data/data_serde.rs
@@ -3,10 +3,13 @@
 use std::borrow::Cow;
 
 use ruma_common::{serde::from_raw_json_value, thirdparty::Medium};
-use serde::{Deserialize, Deserializer, Serialize, de, ser::SerializeStruct};
+use serde::{Deserialize, Deserializer, Serialize, de};
 use serde_json::value::RawValue as RawJsonValue;
 
-use super::{AuthData, CustomThirdPartyId, UserIdentifier};
+use super::{
+    AuthData, CustomThirdPartyUserIdentifier, EmailUserIdentifier, MsisdnUserIdentifier,
+    UserIdentifier,
+};
 
 impl<'de> Deserialize<'de> for AuthData {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
@@ -44,98 +47,96 @@ impl<'de> Deserialize<'de> for AuthData {
     }
 }
 
-impl Serialize for UserIdentifier {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let mut id;
-        match self {
-            Self::UserIdOrLocalpart(user) => {
-                id = serializer.serialize_struct("UserIdentifier", 2)?;
-                id.serialize_field("type", "m.id.user")?;
-                id.serialize_field("user", user)?;
-            }
-            Self::PhoneNumber { country, phone } => {
-                id = serializer.serialize_struct("UserIdentifier", 3)?;
-                id.serialize_field("type", "m.id.phone")?;
-                id.serialize_field("country", country)?;
-                id.serialize_field("phone", phone)?;
-            }
-            Self::Email { address } => {
-                id = serializer.serialize_struct("UserIdentifier", 3)?;
-                id.serialize_field("type", "m.id.thirdparty")?;
-                id.serialize_field("medium", &Medium::Email)?;
-                id.serialize_field("address", address)?;
-            }
-            Self::Msisdn { number } => {
-                id = serializer.serialize_struct("UserIdentifier", 3)?;
-                id.serialize_field("type", "m.id.thirdparty")?;
-                id.serialize_field("medium", &Medium::Msisdn)?;
-                id.serialize_field("address", number)?;
-            }
-            Self::_CustomThirdParty(CustomThirdPartyId { medium, address }) => {
-                id = serializer.serialize_struct("UserIdentifier", 3)?;
-                id.serialize_field("type", "m.id.thirdparty")?;
-                id.serialize_field("medium", &medium)?;
-                id.serialize_field("address", address)?;
-            }
-        }
-        id.end()
-    }
-}
-
 impl<'de> Deserialize<'de> for UserIdentifier {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
         D: Deserializer<'de>,
     {
+        #[derive(Deserialize)]
+        struct ExtractType<'a> {
+            #[serde(borrow, rename = "type")]
+            identifier_type: Cow<'a, str>,
+        }
+
         let json = Box::<RawJsonValue>::deserialize(deserializer)?;
+        let ExtractType { identifier_type } =
+            serde_json::from_str(json.get()).map_err(de::Error::custom)?;
 
-        #[derive(Deserialize)]
-        #[serde(tag = "type")]
-        enum ExtractType {
-            #[serde(rename = "m.id.user")]
-            User,
-            #[serde(rename = "m.id.phone")]
-            Phone,
-            #[serde(rename = "m.id.thirdparty")]
-            ThirdParty,
-        }
-
-        #[derive(Deserialize)]
-        struct UserIdOrLocalpart {
-            user: String,
-        }
-
-        #[derive(Deserialize)]
-        struct ThirdPartyId {
-            medium: Medium,
-            address: String,
-        }
-
-        #[derive(Deserialize)]
-        struct PhoneNumber {
-            country: String,
-            phone: String,
-        }
-
-        let id_type = serde_json::from_str::<ExtractType>(json.get()).map_err(de::Error::custom)?;
-
-        match id_type {
-            ExtractType::User => from_raw_json_value(&json)
-                .map(|user_id: UserIdOrLocalpart| Self::UserIdOrLocalpart(user_id.user)),
-            ExtractType::Phone => from_raw_json_value(&json)
-                .map(|nb: PhoneNumber| Self::PhoneNumber { country: nb.country, phone: nb.phone }),
-            ExtractType::ThirdParty => {
-                let ThirdPartyId { medium, address } = from_raw_json_value(&json)?;
-                match medium {
-                    Medium::Email => Ok(Self::Email { address }),
-                    Medium::Msisdn => Ok(Self::Msisdn { number: address }),
-                    _ => Ok(Self::_CustomThirdParty(CustomThirdPartyId { medium, address })),
+        match identifier_type.as_ref() {
+            "m.id.user" => from_raw_json_value(&json).map(Self::Matrix),
+            "m.id.phone" => from_raw_json_value(&json).map(Self::PhoneNumber),
+            "m.id.thirdparty" => {
+                let id: CustomThirdPartyUserIdentifier = from_raw_json_value(&json)?;
+                match &id.medium {
+                    Medium::Email => Ok(Self::Email(EmailUserIdentifier { address: id.address })),
+                    Medium::Msisdn => Ok(Self::Msisdn(MsisdnUserIdentifier { number: id.address })),
+                    _ => Ok(Self::_CustomThirdParty(id)),
                 }
             }
+            _ => from_raw_json_value(&json).map(Self::_Custom),
         }
+    }
+}
+
+impl Serialize for EmailUserIdentifier {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let Self { address } = self;
+
+        CustomThirdPartyUserIdentifier { medium: Medium::Email, address: address.clone() }
+            .serialize(serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for EmailUserIdentifier {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let CustomThirdPartyUserIdentifier { medium, address } =
+            CustomThirdPartyUserIdentifier::deserialize(deserializer)?;
+
+        if medium != Medium::Email {
+            return Err(de::Error::invalid_value(
+                de::Unexpected::Str(medium.as_str()),
+                &Medium::Email.as_str(),
+            ));
+        }
+
+        Ok(Self { address })
+    }
+}
+
+impl Serialize for MsisdnUserIdentifier {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let Self { number } = self;
+
+        CustomThirdPartyUserIdentifier { medium: Medium::Msisdn, address: number.clone() }
+            .serialize(serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for MsisdnUserIdentifier {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let CustomThirdPartyUserIdentifier { medium, address } =
+            CustomThirdPartyUserIdentifier::deserialize(deserializer)?;
+
+        if medium != Medium::Msisdn {
+            return Err(de::Error::invalid_value(
+                de::Unexpected::Str(medium.as_str()),
+                &Medium::Msisdn.as_str(),
+            ));
+        }
+
+        Ok(Self { number: address })
     }
 }
 
@@ -143,14 +144,17 @@ impl<'de> Deserialize<'de> for UserIdentifier {
 mod tests {
     use assert_matches2::assert_matches;
     use ruma_common::canonical_json::assert_to_canonical_json_eq;
-    use serde_json::{from_value as from_json_value, json};
+    use serde_json::{Value as JsonValue, from_value as from_json_value, json};
 
-    use crate::uiaa::UserIdentifier;
+    use crate::uiaa::{
+        EmailUserIdentifier, MatrixUserIdentifier, MsisdnUserIdentifier, PhoneNumberUserIdentifier,
+        UserIdentifier,
+    };
 
     #[test]
     fn serialize() {
         assert_to_canonical_json_eq!(
-            UserIdentifier::UserIdOrLocalpart("@user:notareal.hs".to_owned()),
+            UserIdentifier::Matrix(MatrixUserIdentifier::new("@user:notareal.hs".to_owned())),
             json!({
                 "type": "m.id.user",
                 "user": "@user:notareal.hs",
@@ -158,10 +162,10 @@ mod tests {
         );
 
         assert_to_canonical_json_eq!(
-            UserIdentifier::PhoneNumber {
-                country: "33".to_owned(),
-                phone: "0102030405".to_owned()
-            },
+            UserIdentifier::PhoneNumber(PhoneNumberUserIdentifier::new(
+                "33".to_owned(),
+                "0102030405".to_owned()
+            )),
             json!({
                 "type": "m.id.phone",
                 "country": "33",
@@ -170,7 +174,7 @@ mod tests {
         );
 
         assert_to_canonical_json_eq!(
-            UserIdentifier::Email { address: "me@myprovider.net".to_owned() },
+            UserIdentifier::Email(EmailUserIdentifier::new("me@myprovider.net".to_owned())),
             json!({
                 "type": "m.id.thirdparty",
                 "medium": "email",
@@ -179,7 +183,7 @@ mod tests {
         );
 
         assert_to_canonical_json_eq!(
-            UserIdentifier::Msisdn { number: "330102030405".to_owned() },
+            UserIdentifier::Msisdn(MsisdnUserIdentifier::new("330102030405".to_owned())),
             json!({
                 "type": "m.id.thirdparty",
                 "medium": "msisdn",
@@ -203,15 +207,18 @@ mod tests {
             "type": "m.id.user",
             "user": "@user:notareal.hs",
         });
-        assert_matches!(from_json_value(json), Ok(UserIdentifier::UserIdOrLocalpart(user)));
-        assert_eq!(user, "@user:notareal.hs");
+        assert_matches!(from_json_value(json), Ok(UserIdentifier::Matrix(id)));
+        assert_eq!(id.user, "@user:notareal.hs");
 
         let json = json!({
             "type": "m.id.phone",
             "country": "33",
             "phone": "0102030405",
         });
-        assert_matches!(from_json_value(json), Ok(UserIdentifier::PhoneNumber { country, phone }));
+        assert_matches!(
+            from_json_value(json),
+            Ok(UserIdentifier::PhoneNumber(PhoneNumberUserIdentifier { country, phone }))
+        );
         assert_eq!(country, "33");
         assert_eq!(phone, "0102030405");
 
@@ -220,16 +227,16 @@ mod tests {
             "medium": "email",
             "address": "me@myprovider.net",
         });
-        assert_matches!(from_json_value(json), Ok(UserIdentifier::Email { address }));
-        assert_eq!(address, "me@myprovider.net");
+        assert_matches!(from_json_value(json), Ok(UserIdentifier::Email(id)));
+        assert_eq!(id.address, "me@myprovider.net");
 
         let json = json!({
             "type": "m.id.thirdparty",
             "medium": "msisdn",
             "address": "330102030405",
         });
-        assert_matches!(from_json_value(json), Ok(UserIdentifier::Msisdn { number }));
-        assert_eq!(number, "330102030405");
+        assert_matches!(from_json_value(json), Ok(UserIdentifier::Msisdn(id)));
+        assert_eq!(id.number, "330102030405");
 
         let json = json!({
             "type": "m.id.thirdparty",
@@ -240,5 +247,21 @@ mod tests {
         let (medium, address) = id.as_third_party_id().unwrap();
         assert_eq!(medium.as_str(), "robot");
         assert_eq!(address, "01110010");
+    }
+
+    #[test]
+    fn custom_identifier_roundtrip() {
+        let json = json!({
+            "type": "local.dev.identifier",
+            "foo": "bar",
+        });
+
+        let id = from_json_value::<UserIdentifier>(json.clone()).unwrap();
+        assert_eq!(id.identifier_type(), "local.dev.identifier");
+        assert_matches!(id.custom_identifier_data(), Some(data));
+        assert_matches!(data.get("foo"), Some(JsonValue::String(foo)));
+        assert_eq!(foo, "bar");
+
+        assert_to_canonical_json_eq!(id, json);
     }
 }