+++ crates/ruma-client-api/CHANGELOG.md
@@ -11,6 +11,9 @@ Breaking changes:
   spec.
 - `BackupAlgorithm::MegolmBackupV1Curve25519AesSha2` is now a tuple variant
   containing a non-exhaustive struct.
+- `UserIdentifier` can now represent an identifier with a custom type, and its
+  variants were changed to tuple variants containing a non-exhaustive struct.
+  - `UserIdentifier::UserIdOrLocalpart` was renamed to `UserIdentifier::Matrix`.
 
 Improvements: